PSR-4 support. Capitalized namespace. Moved helper class to Jyggen\Curl.
