+++ composer.json
@@ -23,7 +23,10 @@
         "league/phpunit-coverage-listener": "~1.1"
     },
     "autoload": {
-        "psr-0": { "jyggen\\Curl": "src/" }
+        "psr-4": {
+            "Jyggen\\Curl\\": "src",
+            "Jyggen\\Curl\\Tests\\": "tests"
+        }
     },
     "extra": {
         "branch-alias": {
